+++ packages/@coorpacademy-components/src/template/app-review/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import Loader from '../../atom/loader';
 import Onboarding from './onboarding';
 import Skills from './skills';
-import Slides from './slides';
+import Player from './player';
 import propTypes, {ViewNames} from './prop-types';
 
 // -----------------------------------------------------------------------------
@@ -14,7 +14,7 @@ const AppReview = ({viewName, onboarding, skills, slides, ...dispatchers}) => {
     case ViewNames.onboarding:
       return <Onboarding {...onboarding} {...dispatchers} />;
     case ViewNames.slides:
-      return <Slides {...slides} {...dispatchers} />;
+      return <Player {...slides} {...dispatchers} />;
     default:
       return <Loader />;
   }
