+++ packages/@coorpacademy-components/src/template/app-review/player/index.js
