+++ packages/@coorpacademy-components/src/template/app-review/player/prop-types.js
