+++ packages/@coorpacademy-components/src/template/app-review/player/stacked-slides.js
