+++ packages/@coorpacademy-components/src/template/app-review/player/style.css
