+++ packages/@coorpacademy-components/src/template/app-review/player/test/fixtures/all-ok.js
