+++ packages/@coorpacademy-components/src/template/app-review/player/test/slides-review.js
